refactor: replace components to correct folders
