+++ client/src/components/Dashboard/LearnSection/Recruitment/components/Recruitment/RecruitmentStats.component.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { motion } from 'framer-motion';
 import { FaUsers } from 'react-icons/fa';
-import type { RecruitmentStats } from '../types/recruitment.types';
+import type { RecruitmentStats } from '../../types/recruitment.types';
 
 type RecruitmentStatsProps = {
   stats: RecruitmentStats;
