+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Common/InterviewBlockCard.component.tsx
