+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Common/InterviewTipsSection.component.tsx
