+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Results/components/RecommendationsCard.component.tsx
