+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Results/components/StatsCard.component.tsx
