+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Results/components/index.ts
@@ -1,0 +1,7 @@
+export { ResultsHeader } from './ResultsHeader.component';
+export { MainResultCard } from './MainResultCard.component';
+export { StatsCard } from './StatsCard.component';
+export { CategoryStatsCard } from './CategoryStatsCard.component';
+export { RecommendationsCard } from './RecommendationsCard.component';
+export { DetailedAnswers } from './DetailedAnswers.component';
+export { ResultsActions } from './ResultsActions.component';
