+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Theory/StartButton.component.tsx
