+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Theory/TestInfoSection.component.tsx
