+++ client/src/components/Dashboard/LearnSection/Recruitment/components/TechnicalInterview/components/Theory/TheoryHeader.component.tsx
